requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
aiosmtplib==3.0.2
alaric==1.4.0
astunparse==1.6.3
attrs==24.2.0
braceexpand==0.1.7
click==8.1.7
colorama==0.4.6
discord-ext-menus==1.1
discord.py==2.4.0
dnspython==2.3.0
frozenlist==1.4.1
idna==3.10
import_expression==1.1.5
jishaku==2.5.2
motor==3.6.0
multidict==6.1.0
orjson==3.10.7
parsedatetime==2.6
pycryptodome==3.17
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
six==1.16.0
typing_extensions==4.12.2
yarl==1.13.1
